LPS-150725 Splitting into files
